+++ packages/user/src/components/event/racing/dashboard/card/index.tsx
@@ -1,6 +1,6 @@
 import TriggerButtonWrapper from 'src/components/common/TriggerButtonWrapper.tsx';
+import TeamSelectModal from 'src/components/shared/modal/teamSelectModal/index.tsx';
 import useAuth from 'src/hooks/useAuth.tsx';
-import TeamSelectModal from './teamSelectModal/index.tsx';
 import RacingTeamCard from './RacingTeamCard.tsx';
 import UnassignedCard from './UnassignedCard.tsx';
