+++ packages/user/src/components/shared/modal/teamSelectModal/index.tsx
@@ -1,0 +1,15 @@
+import Modal, { ModalProps } from 'src/components/common/Modal.tsx';
+import LoginModal from 'src/components/shared/modal/login/index.tsx';
+import useAuth from 'src/hooks/useAuth.tsx';
+
+interface TeamSelectModalProps extends Omit<ModalProps, 'children'> {}
+
+export default function TeamSelectModal({ openTrigger }: TeamSelectModalProps) {
+	const { user } = useAuth();
+
+	if (!user) {
+		return <LoginModal openTrigger={openTrigger} />;
+	}
+
+	return <Modal openTrigger={openTrigger}>유형 검사</Modal>;
+}
